Refactor CLI app structure
* services are managed by the ServiceContainer
* the app is bootstrapped by Main
