+++ apps/cli/src/commands/serve.command.ts
@@ -11,9 +11,9 @@ import { ConfirmCommand } from "../admin-console/commands/confirm.command";
 import { ShareCommand } from "../admin-console/commands/share.command";
 import { LockCommand } from "../auth/commands/lock.command";
 import { UnlockCommand } from "../auth/commands/unlock.command";
-import { Main } from "../bw";
 import { Response } from "../models/response";
 import { FileResponse } from "../models/response/file.response";
+import { ServiceContainer } from "../service-container";
 import { GenerateCommand } from "../tools/generate.command";
 import {
   SendEditCommand,
@@ -55,116 +55,119 @@ export class ServeCommand {
   private sendListCommand: SendListCommand;
   private sendRemovePasswordCommand: SendRemovePasswordCommand;
 
-  constructor(protected main: Main) {
+  constructor(protected serviceContainer: ServiceContainer) {
     this.getCommand = new GetCommand(
-      this.main.cipherService,
-      this.main.folderService,
-      this.main.collectionService,
-      this.main.totpService,
-      this.main.auditService,
-      this.main.cryptoService,
-      this.main.stateService,
-      this.main.searchService,
-      this.main.apiService,
-      this.main.organizationService,
-      this.main.eventCollectionService,
-      this.main.billingAccountProfileStateService,
+      this.serviceContainer.cipherService,
+      this.serviceContainer.folderService,
+      this.serviceContainer.collectionService,
+      this.serviceContainer.totpService,
+      this.serviceContainer.auditService,
+      this.serviceContainer.cryptoService,
+      this.serviceContainer.stateService,
+      this.serviceContainer.searchService,
+      this.serviceContainer.apiService,
+      this.serviceContainer.organizationService,
+      this.serviceContainer.eventCollectionService,
+      this.serviceContainer.billingAccountProfileStateService,
     );
     this.listCommand = new ListCommand(
-      this.main.cipherService,
-      this.main.folderService,
-      this.main.collectionService,
-      this.main.organizationService,
-      this.main.searchService,
-      this.main.organizationUserService,
-      this.main.apiService,
-      this.main.eventCollectionService,
+      this.serviceContainer.cipherService,
+      this.serviceContainer.folderService,
+      this.serviceContainer.collectionService,
+      this.serviceContainer.organizationService,
+      this.serviceContainer.searchService,
+      this.serviceContainer.organizationUserService,
+      this.serviceContainer.apiService,
+      this.serviceContainer.eventCollectionService,
     );
     this.createCommand = new CreateCommand(
-      this.main.cipherService,
-      this.main.folderService,
-      this.main.cryptoService,
-      this.main.apiService,
-      this.main.folderApiService,
-      this.main.billingAccountProfileStateService,
+      this.serviceContainer.cipherService,
+      this.serviceContainer.folderService,
+      this.serviceContainer.cryptoService,
+      this.serviceContainer.apiService,
+      this.serviceContainer.folderApiService,
+      this.serviceContainer.billingAccountProfileStateService,
     );
     this.editCommand = new EditCommand(
-      this.main.cipherService,
-      this.main.folderService,
-      this.main.cryptoService,
-      this.main.apiService,
-      this.main.folderApiService,
+      this.serviceContainer.cipherService,
+      this.serviceContainer.folderService,
+      this.serviceContainer.cryptoService,
+      this.serviceContainer.apiService,
+      this.serviceContainer.folderApiService,
     );
     this.generateCommand = new GenerateCommand(
-      this.main.passwordGenerationService,
-      this.main.stateService,
+      this.serviceContainer.passwordGenerationService,
+      this.serviceContainer.stateService,
     );
-    this.syncCommand = new SyncCommand(this.main.syncService);
+    this.syncCommand = new SyncCommand(this.serviceContainer.syncService);
     this.statusCommand = new StatusCommand(
-      this.main.environmentService,
-      this.main.syncService,
-      this.main.stateService,
-      this.main.authService,
+      this.serviceContainer.environmentService,
+      this.serviceContainer.syncService,
+      this.serviceContainer.stateService,
+      this.serviceContainer.authService,
     );
     this.deleteCommand = new DeleteCommand(
-      this.main.cipherService,
-      this.main.folderService,
-      this.main.apiService,
-      this.main.folderApiService,
-      this.main.billingAccountProfileStateService,
+      this.serviceContainer.cipherService,
+      this.serviceContainer.folderService,
+      this.serviceContainer.apiService,
+      this.serviceContainer.folderApiService,
+      this.serviceContainer.billingAccountProfileStateService,
     );
     this.confirmCommand = new ConfirmCommand(
-      this.main.apiService,
-      this.main.cryptoService,
-      this.main.organizationUserService,
+      this.serviceContainer.apiService,
+      this.serviceContainer.cryptoService,
+      this.serviceContainer.organizationUserService,
     );
-    this.restoreCommand = new RestoreCommand(this.main.cipherService);
-    this.shareCommand = new ShareCommand(this.main.cipherService);
-    this.lockCommand = new LockCommand(this.main.vaultTimeoutService);
+    this.restoreCommand = new RestoreCommand(this.serviceContainer.cipherService);
+    this.shareCommand = new ShareCommand(this.serviceContainer.cipherService);
+    this.lockCommand = new LockCommand(this.serviceContainer.vaultTimeoutService);
     this.unlockCommand = new UnlockCommand(
-      this.main.accountService,
-      this.main.masterPasswordService,
-      this.main.cryptoService,
-      this.main.stateService,
-      this.main.cryptoFunctionService,
-      this.main.apiService,
-      this.main.logService,
-      this.main.keyConnectorService,
-      this.main.environmentService,
-      this.main.syncService,
-      this.main.organizationApiService,
-      async () => await this.main.logout(),
-      this.main.kdfConfigService,
+      this.serviceContainer.accountService,
+      this.serviceContainer.masterPasswordService,
+      this.serviceContainer.cryptoService,
+      this.serviceContainer.stateService,
+      this.serviceContainer.cryptoFunctionService,
+      this.serviceContainer.apiService,
+      this.serviceContainer.logService,
+      this.serviceContainer.keyConnectorService,
+      this.serviceContainer.environmentService,
+      this.serviceContainer.syncService,
+      this.serviceContainer.organizationApiService,
+      async () => await this.serviceContainer.logout(),
+      this.serviceContainer.kdfConfigService,
     );
 
     this.sendCreateCommand = new SendCreateCommand(
-      this.main.sendService,
-      this.main.environmentService,
-      this.main.sendApiService,
-      this.main.billingAccountProfileStateService,
+      this.serviceContainer.sendService,
+      this.serviceContainer.environmentService,
+      this.serviceContainer.sendApiService,
+      this.serviceContainer.billingAccountProfileStateService,
     );
-    this.sendDeleteCommand = new SendDeleteCommand(this.main.sendService, this.main.sendApiService);
+    this.sendDeleteCommand = new SendDeleteCommand(
+      this.serviceContainer.sendService,
+      this.serviceContainer.sendApiService,
+    );
     this.sendGetCommand = new SendGetCommand(
-      this.main.sendService,
-      this.main.environmentService,
-      this.main.searchService,
-      this.main.cryptoService,
+      this.serviceContainer.sendService,
+      this.serviceContainer.environmentService,
+      this.serviceContainer.searchService,
+      this.serviceContainer.cryptoService,
     );
     this.sendEditCommand = new SendEditCommand(
-      this.main.sendService,
+      this.serviceContainer.sendService,
       this.sendGetCommand,
-      this.main.sendApiService,
-      this.main.billingAccountProfileStateService,
+      this.serviceContainer.sendApiService,
+      this.serviceContainer.billingAccountProfileStateService,
     );
     this.sendListCommand = new SendListCommand(
-      this.main.sendService,
-      this.main.environmentService,
-      this.main.searchService,
+      this.serviceContainer.sendService,
+      this.serviceContainer.environmentService,
+      this.serviceContainer.searchService,
     );
     this.sendRemovePasswordCommand = new SendRemovePasswordCommand(
-      this.main.sendService,
-      this.main.sendApiService,
-      this.main.environmentService,
+      this.serviceContainer.sendService,
+      this.serviceContainer.sendApiService,
+      this.serviceContainer.environmentService,
     );
   }
 
@@ -172,7 +175,7 @@ export class ServeCommand {
     const protectOrigin = !options.disableOriginProtection;
     const port = options.port || 8087;
     const hostname = options.hostname || "localhost";
-    this.main.logService.info(
+    this.serviceContainer.logService.info(
       `Starting server on ${hostname}:${port} with ${
         protectOrigin ? "origin protection" : "no origin protection"
       }`,
@@ -187,7 +190,7 @@ export class ServeCommand {
       .use(async (ctx, next) => {
         if (protectOrigin && ctx.headers.origin != undefined) {
           ctx.status = 403;
-          this.main.logService.warning(
+          this.serviceContainer.logService.warning(
             `Blocking request from "${
               Utils.isNullOrEmpty(ctx.headers.origin)
                 ? "(Origin header value missing)"
@@ -407,7 +410,7 @@ export class ServeCommand {
       .use(router.routes())
       .use(router.allowedMethods())
       .listen(port, hostname === "all" ? null : hostname, () => {
-        this.main.logService.info("Listening on " + hostname + ":" + port);
+        this.serviceContainer.logService.info("Listening on " + hostname + ":" + port);
       });
   }
 
@@ -426,12 +429,12 @@ export class ServeCommand {
   }
 
   private async errorIfLocked(res: koa.Response) {
-    const authed = await this.main.stateService.getIsAuthenticated();
+    const authed = await this.serviceContainer.stateService.getIsAuthenticated();
     if (!authed) {
       this.processResponse(res, Response.error("You are not logged in."));
       return true;
     }
-    if (await this.main.cryptoService.hasUserKey()) {
+    if (await this.serviceContainer.cryptoService.hasUserKey()) {
       return false;
     }
     this.processResponse(res, Response.error("Vault is locked."));
