+++ apps/cli/src/oss-program-definitions.ts
@@ -1,0 +1,5 @@
+import { Program } from "./program";
+import { SendProgram } from "./tools/send/send.program";
+import { VaultProgram } from "./vault.program";
+
+export const ossProgramDefinitions = [Program, VaultProgram, SendProgram];
