+++ apps/cli/src/vault.program.ts
@@ -2,7 +2,6 @@ import { program, Command } from "commander";
 
 import { ConfirmCommand } from "./admin-console/commands/confirm.command";
 import { ShareCommand } from "./admin-console/commands/share.command";
-import { Main } from "./bw";
 import { EditCommand } from "./commands/edit.command";
 import { GetCommand } from "./commands/get.command";
 import { ListCommand } from "./commands/list.command";
@@ -18,10 +17,6 @@ import { DeleteCommand } from "./vault/delete.command";
 const writeLn = CliUtils.writeLn;
 
 export class VaultProgram extends Program {
-  constructor(protected main: Main) {
-    super(main);
-  }
-
   async register() {
     program
       .addCommand(this.listCommand())
@@ -108,14 +103,14 @@ export class VaultProgram extends Program {
 
         await this.exitIfLocked();
         const command = new ListCommand(
-          this.main.cipherService,
-          this.main.folderService,
-          this.main.collectionService,
-          this.main.organizationService,
-          this.main.searchService,
-          this.main.organizationUserService,
-          this.main.apiService,
-          this.main.eventCollectionService,
+          this.serviceContainer.cipherService,
+          this.serviceContainer.folderService,
+          this.serviceContainer.collectionService,
+          this.serviceContainer.organizationService,
+          this.serviceContainer.searchService,
+          this.serviceContainer.organizationUserService,
+          this.serviceContainer.apiService,
+          this.serviceContainer.eventCollectionService,
         );
         const response = await command.run(object, cmd);
 
@@ -177,18 +172,18 @@ export class VaultProgram extends Program {
 
         await this.exitIfLocked();
         const command = new GetCommand(
-          this.main.cipherService,
-          this.main.folderService,
-          this.main.collectionService,
-          this.main.totpService,
-          this.main.auditService,
-          this.main.cryptoService,
-          this.main.stateService,
-          this.main.searchService,
-          this.main.apiService,
-          this.main.organizationService,
-          this.main.eventCollectionService,
-          this.main.billingAccountProfileStateService,
+          this.serviceContainer.cipherService,
+          this.serviceContainer.folderService,
+          this.serviceContainer.collectionService,
+          this.serviceContainer.totpService,
+          this.serviceContainer.auditService,
+          this.serviceContainer.cryptoService,
+          this.serviceContainer.stateService,
+          this.serviceContainer.searchService,
+          this.serviceContainer.apiService,
+          this.serviceContainer.organizationService,
+          this.serviceContainer.eventCollectionService,
+          this.serviceContainer.billingAccountProfileStateService,
         );
         const response = await command.run(object, id, cmd);
         this.processResponse(response);
@@ -225,12 +220,12 @@ export class VaultProgram extends Program {
 
         await this.exitIfLocked();
         const command = new CreateCommand(
-          this.main.cipherService,
-          this.main.folderService,
-          this.main.cryptoService,
-          this.main.apiService,
-          this.main.folderApiService,
-          this.main.billingAccountProfileStateService,
+          this.serviceContainer.cipherService,
+          this.serviceContainer.folderService,
+          this.serviceContainer.cryptoService,
+          this.serviceContainer.apiService,
+          this.serviceContainer.folderApiService,
+          this.serviceContainer.billingAccountProfileStateService,
         );
         const response = await command.run(object, encodedJson, cmd);
         this.processResponse(response);
@@ -271,11 +266,11 @@ export class VaultProgram extends Program {
 
         await this.exitIfLocked();
         const command = new EditCommand(
-          this.main.cipherService,
-          this.main.folderService,
-          this.main.cryptoService,
-          this.main.apiService,
-          this.main.folderApiService,
+          this.serviceContainer.cipherService,
+          this.serviceContainer.folderService,
+          this.serviceContainer.cryptoService,
+          this.serviceContainer.apiService,
+          this.serviceContainer.folderApiService,
         );
         const response = await command.run(object, id, encodedJson, cmd);
         this.processResponse(response);
@@ -312,11 +307,11 @@ export class VaultProgram extends Program {
 
         await this.exitIfLocked();
         const command = new DeleteCommand(
-          this.main.cipherService,
-          this.main.folderService,
-          this.main.apiService,
-          this.main.folderApiService,
-          this.main.billingAccountProfileStateService,
+          this.serviceContainer.cipherService,
+          this.serviceContainer.folderService,
+          this.serviceContainer.apiService,
+          this.serviceContainer.folderApiService,
+          this.serviceContainer.billingAccountProfileStateService,
         );
         const response = await command.run(object, id, cmd);
         this.processResponse(response);
@@ -341,7 +336,7 @@ export class VaultProgram extends Program {
         }
 
         await this.exitIfLocked();
-        const command = new RestoreCommand(this.main.cipherService);
+        const command = new RestoreCommand(this.serviceContainer.cipherService);
         const response = await command.run(object, id);
         this.processResponse(response);
       });
@@ -379,7 +374,7 @@ export class VaultProgram extends Program {
       })
       .action(async (id, organizationId, encodedJson, cmd) => {
         await this.exitIfLocked();
-        const command = new ShareCommand(this.main.cipherService);
+        const command = new ShareCommand(this.serviceContainer.cipherService);
         const response = await command.run(id, organizationId, encodedJson);
         this.processResponse(response);
       });
@@ -408,9 +403,9 @@ export class VaultProgram extends Program {
 
         await this.exitIfLocked();
         const command = new ConfirmCommand(
-          this.main.apiService,
-          this.main.cryptoService,
-          this.main.organizationUserService,
+          this.serviceContainer.apiService,
+          this.serviceContainer.cryptoService,
+          this.serviceContainer.organizationUserService,
         );
         const response = await command.run(object, id, cmd);
         this.processResponse(response);
@@ -437,9 +432,9 @@ export class VaultProgram extends Program {
       .action(async (format, filepath, options) => {
         await this.exitIfLocked();
         const command = new ImportCommand(
-          this.main.importService,
-          this.main.organizationService,
-          this.main.syncService,
+          this.serviceContainer.importService,
+          this.serviceContainer.organizationService,
+          this.serviceContainer.syncService,
         );
         const response = await command.run(format, filepath, options);
         this.processResponse(response);
@@ -484,9 +479,9 @@ export class VaultProgram extends Program {
       .action(async (options) => {
         await this.exitIfLocked();
         const command = new ExportCommand(
-          this.main.exportService,
-          this.main.policyService,
-          this.main.eventCollectionService,
+          this.serviceContainer.exportService,
+          this.serviceContainer.policyService,
+          this.serviceContainer.eventCollectionService,
         );
         const response = await command.run(options);
         this.processResponse(response);
